refactor: Update date formatting and enhance snapshot data handling
- Changed date formatting in InvestableTokens and proposal utility functions to use numeric month representation for consistency.
- Updated DashboardCharts and FundCharts components to utilize new snapshot data structure, improving data accuracy and clarity.
- Enhanced interval type handling in various components to support daily, weekly, and monthly views for better user experience.
- Refactored hooks and queries to align with updated data fields, ensuring accurate data fetching and representation across the application.

+++ app/dashboard/fund/components/InvestableTokens.tsx
@@ -61,7 +61,7 @@ export function InvestableTokens({ network, setActiveTab, selectedNetwork = 'eth
     const date = new Date(parseInt(timestamp) * 1000)
     return date.toLocaleDateString('en-US', {
       year: 'numeric',
-      month: 'short',
+      month: 'numeric',
       day: 'numeric'
     })
   }
