+++ app/vote/utils/proposal.ts
@@ -327,7 +327,7 @@ export const formatDate = (date: Date): string => {
   
   return date.toLocaleDateString('en-US', {
     year: 'numeric',
-    month: 'short',
+    month: 'numeric',
     day: 'numeric',
     hour: '2-digit',
     minute: '2-digit'
